refactor: split skills pool capability persistence
